HomeWork4/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
python-jose
httpx
pyodbc
